Rework hidden frame handling to fix interlaced files
Commit 1/2 to fix interlaced files. After this commit,
seeking is correct on interlaced h264 with threads=1, but
not yet with more.

Remove the PAFF adjustment hacks, which don't work for actual
mixed interlaced/progressive content. The second field of each
frame is already being marked as hidden during indexing, but
previously that would not actually be used during decoding
since DecodePacket checks the AV_PKG_FLAG_DISCARD flag rather
than ffms2's own Hidden flag.

Hence, this commit makes DecodePacket check the Hidden flag instead.
It also simplifies the DelayCounter logic a bit by simply increasing
the DelayCounter every time a non-hidden packet is sent and decreasing
it every time a frame is received.

As far as I can see, the added AllowHidden arguments for
FrameFromPos and FrameFromPTS are not actually necessary, and can just
always be true, but this way less existing behavior is changed. This
can be revisited when a sample comes up where this would actually
make a difference (i.e. the first frame after seeking being hidden,
or ending up somewhere unknown after a seek in a file with hidden
frames).