Fix hidden frames not being skipped right after seek
An example file where this is necessary is ffms2_seeking_issue.mp4
in the doom9 collection of seek test sample files.

This solution is quite ugly but it's the best I could find.
The only alternative I can think of is seeking to the same point
twice on every seek, once to compute CurrentFrame and a second time
to actually start the decoding loop. This might make the GetFrame
loop a lot cleaner but it'd be a bigger refactor and make seeking
slower again.