Force a seek right after initializing source filter
The Seek(0) in the FFMS_VideoSource constructor can break
the invariants assumed during the GetFrame loop later on when
the first frame in output order is not the first frame in decoding
order. An example for this is 3d-parrot.MTS from the doom9 collection.
We could try to ensure that this initial seek behaves just like the
future seeks in the GetFrame loop do, but it's simpler to just force
a seek at the start to go into decoding with a clean slate.

Similarly, don't output the frame we decoded in the constructor,
just to be sure.